Fixed Memory leak in Chat system Now business list is loaded as map is
My last minute changes before that last commit caused chat messages to stop working.  I reverted those changes to the last version.

+++ MainSourceCode/ChatMessagesViewController.m
@@ -101,9 +101,8 @@
 #pragma mark -
 #pragma mark UITableViewDataSource
 
-- (NSInteger)tableView:(UITableView *)tableView numberOfRowsInSection:(NSInteger)section {
+- (int)tableView:(UITableView *)tableView numberOfRowsInSection:(NSInteger)section {
     return [DataModel sharedDataModelManager].messages.count;
-
 }
 
 - (UITableViewCell *)tableView:(UITableView *)tableView cellForRowAtIndexPath:(NSIndexPath *)indexPath {
